emu86: Fix escape, up arrow, and down arrow causing a freeze
It turns out that there seems to be a bug in recent glibc where
a terminal set from blocking to unlblocking and back again doesn't
work - all reads are stuck in unlocking mode. Switching over to
Linux read() instead of getchar() allows use to read escape sequences
fully without getting stuck in non-blocking mode.

I'm not sure when this issue started, but I think it worked back in
2015 with the BYU CAEDM Red hat machines. I bet that recent Linux
versions have a newer glibc that messed this up.